feat(cli): Update cli with configurations

+++ src/index.ts
@@ -6,6 +6,7 @@ import * as Handlebars from "handlebars";
 export interface Args {
   source: string
   target: string
+  config?: string
 }
 
 function removeTemplateDir(templateDir: string) {
@@ -17,12 +18,13 @@ function removeTemplateDir(templateDir: string) {
   fs.mkdirSync(templateDir)
 }
 
-function generatePackageJson(packageJson, templateDir: string, name: string) {
+function generatePackageJson(packageJson, templateDir: string, name: string, templateConfig: TemplateConfig) {
   let packageJsonContent = JSON.stringify({
     name: name,
     version: packageJson.version,
-    keywords: [...(packageJson.keywords || []), "cra-template", "react"],
-    description: packageJson.description,
+    keywords: [...(templateConfig.keywords ?? [...(packageJson.keywords || [])]), "cra-template", "react"],
+    description: templateConfig.description ?? packageJson.description,
+    private: templateConfig.private,
     files: [
       "template",
       "template.json"
@@ -56,32 +58,40 @@ function generateTemplateJson(packageJson, templateDir: string) {
   fs.writeFileSync(path.join(templateDir, 'template.json'), templateJson)
 }
 
-function addTemplateResources(templateDir: string, source: string) {
+function addTemplateResources(templateDir: string, source: string, templateConfig: TemplateConfig) {
   fs.mkdirSync(path.join(templateDir, 'template'))
 
-  for (let name of fs.readdirSync(source)) {
-    if (!['package.json', 'package-lock.json', 'yarn.lock', 'npm.lock', '.gitignore', 'node_modules'].includes(name)) {
-      fse.copySync(path.join(source, name), path.join(templateDir, 'template', name), {
-        recursive: true
-      })
+  if (!!templateConfig.include) {
+    for (let src of templateConfig.include) {
+      if (!!fs.existsSync(path.join(source, src))) {
+        fse.copySync(path.join(source, src), path.join(templateDir, 'template', path.basename(src)), { recursive: true })
+      }
     }
-    if (name === '.gitignore') {
-      fse.copySync(path.join(source, name), path.join(templateDir, 'template', 'gitignore'), {
-        recursive: true
-      })
+  }
+
+  fse.copySync(source, path.join(templateDir, 'template'), {
+    recursive: true,
+    filter(src: string, dest: string) {
+      for (let skipPath of [...templateSkipPaths, ...(templateConfig.exclude ?? []), ...(templateConfig.readme ? [templateConfig.readme] : [])]) {
+        if (src.startsWith(path.join(source, skipPath))) return false
+      }
+      return true
     }
+  })
+  if (fs.existsSync(path.join(source, '.gitignore'))) {
+    fse.copySync(path.join(source, '.gitignore'), path.join(templateDir, 'template', 'gitignore'))
   }
 }
 
-function extractTemplateName(packageJson) {
-  let groups = packageJson.name.match(/^@([^/]+)\/([^/]+$)/);
+function extractTemplateName(templateConfig: TemplateConfig) {
+  let groups = templateConfig.name.match(/^@([^/]+)\/([^/]+$)/);
   if (!!groups) {
     return [`cra-template-${groups[2]}`, groups[1]];
   }
-  return [`cra-template-${packageJson.name}`];
+  return [`cra-template-${templateConfig.name}`];
 }
 
-export async function templetize({target, source}: Args) {
+export async function templetize({target, source, config}: Args) {
   if(!fs.existsSync(source)) {
     console.error("Cannot find source directory")
     process.exit(1)
@@ -98,19 +108,32 @@ export async function templetize({target, source}: Args) {
     process.exit(1)
   }
 
+  if (config && !fs.existsSync(config)) {
+    console.error("Invaid configuration")
+    process.exit(1)
+  }
+
   let packageJson = JSON.parse(fs.readFileSync(packageJsonPath).toString());
-  let [templateName, scope] = extractTemplateName(packageJson);
+
+  const templateConfig: TemplateConfig = config ? JSON.parse(fs.readFileSync(config).toString()) : {
+    name: packageJson.name
+  }
+  let [templateName, scope] = extractTemplateName(templateConfig);
   let templateDir = path.join(target, templateName);
 
   removeTemplateDir(templateDir);
-  generatePackageJson(packageJson, templateDir, scope ? `@${scope}/${templateName}`: templateName);
+  generatePackageJson(packageJson, templateDir, scope ? `@${scope}/${templateName}`: templateName, templateConfig);
   generateTemplateJson(packageJson, templateDir);
-  addTemplateResources(templateDir, source);
+  addTemplateResources(templateDir, source, templateConfig);
 
-  let readmeContent = Handlebars.compile(README_TEMPLATE)({
-    ...packageJson
-  });
-  fs.writeFileSync(path.join(templateDir, 'README.md'), readmeContent)
+  if (templateConfig.readme && fs.existsSync(path.join(path.dirname(config), templateConfig.readme))) {
+    fse.copySync(path.join(path.dirname(config), templateConfig.readme), path.join(templateDir, 'README.md'))
+  } else {
+    let readmeContent = Handlebars.compile(README_TEMPLATE)({
+      ...packageJson
+    });
+    fs.writeFileSync(path.join(templateDir, 'README.md'), readmeContent)
+  }
 }
 
 const README_TEMPLATE = `
@@ -125,3 +148,17 @@ ${"```"}shell
 npx create-react-app <project-name> --template {{name}}
 ${"```"}
 `
+
+interface TemplateConfig {
+  name?: string,
+  include?: string[]
+  exclude?: string[]
+  readme?: string
+  keywords?: string[]
+  private?: boolean
+  description?: string
+}
+
+const templateSkipPaths = [
+  'package.json', 'package-lock.json', 'yarn.lock', 'npm.lock', '.gitignore', 'node_modules', 'dist', 'template-config.json'
+]
